Standalone stablehlo compiler
